refactored moved functions to other files to make it cleaner
