+++ src/Finder.py
@@ -1,4 +1,4 @@
-import Exception as ex
+from src import Exception as ex
 from PBSclasses import EncounterMethod, Move, Item, Species
 from PBSclasses.TrainerTypes import TrainerType
