+++ src/main.py
@@ -1,38 +1,26 @@
-from Finder import get_species_from_name, get_trainer_type_from_name, get_item_from_name
+from pathlib import Path
+
+from src.Finder import get_species_from_name, get_trainer_type_from_name, get_item_from_name
 from PBSclasses.Environment import Environment
 
-
-import FileLoader
-from Exporter import (
-    deserialize_shadow,
-    deserialize_phone,
-    deserialize_species,
-    deserialize_type,
-    deserialize_ability,
-    deserialize_move,
-    deserialize_trainer_types,
-    deserialize_item,
-    deserialize_connection,
-    deserialize_metadata,
-    deserialize_townmap,
-    deserialize_encounter,
-    deserialize_trainer,
-)
+from src import FileLoader
+from src.Exporter import deserialize_encounter
 
 if __name__ == "__main__":
-    csv_trainer = FileLoader.file_csv_tolist("PBS/trainers.txt")
-    csv_trainer_type = FileLoader.file_csv_tolist("PBS/trainertypes.txt")
-    csv_move = FileLoader.file_csv_tolist("PBS/moves.txt")
-    csv_item = FileLoader.file_csv_tolist("PBS/items.txt")
-    csv_ability = FileLoader.file_csv_tolist("PBS/abilities.txt")
-    csv_encounter = FileLoader.file_csv_tolist("PBS/encounters.txt")
-    csv_connection = FileLoader.file_csv_tolist("PBS/connections.txt")
-    equal_pokemon_species = FileLoader.file_equal_to_list("PBS/pokemon.txt")
-    equal_pokemon_shadow = FileLoader.file_equal_to_list("PBS/shadowmoves.txt")
-    equal_phone = FileLoader.file_equal_to_list("PBS/phone.txt")
-    equal_type = FileLoader.file_equal_to_list("PBS/types.txt")
-    equal_townmap = FileLoader.file_equal_to_list("PBS/townmap.txt")
-    equal_metadata = FileLoader.file_equal_to_list("PBS/metadata.txt")
+    pbs_location = "./PBS/"
+    csv_trainer = FileLoader.file_csv_tolist(pbs_location + "trainers.txt")
+    csv_trainer_type = FileLoader.file_csv_tolist(pbs_location + "trainertypes.txt")
+    csv_move = FileLoader.file_csv_tolist(pbs_location + "moves.txt")
+    csv_item = FileLoader.file_csv_tolist(pbs_location + "items.txt")
+    csv_ability = FileLoader.file_csv_tolist(pbs_location + "abilities.txt")
+    csv_encounter = FileLoader.file_csv_tolist(pbs_location + "encounters.txt")
+    csv_connection = FileLoader.file_csv_tolist(pbs_location + "connections.txt")
+    equal_pokemon_species = FileLoader.file_equal_to_list(pbs_location + "pokemon.txt")
+    equal_pokemon_shadow = FileLoader.file_equal_to_list(pbs_location + "shadowmoves.txt")
+    equal_phone = FileLoader.file_equal_to_list(pbs_location + "phone.txt")
+    equal_type = FileLoader.file_equal_to_list(pbs_location + "types.txt")
+    equal_townmap = FileLoader.file_equal_to_list(pbs_location + "townmap.txt")
+    equal_metadata = FileLoader.file_equal_to_list(pbs_location + "metadata.txt")
 
     env = Environment()
     env.load_environment(
